requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
cssselect==1.1.0
Django==3.2.9
django-crispy-forms==1.13.0
django-tinymce4-lite==1.8.0
jsmin==3.0.0
lxml==4.6.4
markdown2==2.4.1
Pillow==8.4.0
pyquery==1.4.3
pytz==2021.3
readtime==1.1.1
soupsieve==2.3
sqlparse==0.4.2
